第五章/Code_Data/requirements.txt
selenium>=4.8.0
webdriver-manager>=3.8.5
pytest>=7.0
requests>=2.28
pytest-html>=3.1
pytest-cov>=4.0
jira>=3.5
TestLink-API-Python-client>=0.8
